agentengine/pyproject.toml
[project]
name = "snowflake_adk_agent"
version = "0.1.0"
description = "JIRA MCP Snowflake Agent for AgentSpace deployment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform[agent_engines,adk]>=1.101.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]
